Code seperated to files. Monthly recurring and monthly expenditure comparison functions completed
