+++ category.html
@@ -1,0 +1,81 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Spend Wise</title>
+        <link rel="stylesheet" href="./css/material.min.css">
+        <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+        <link rel="stylesheet" href="css/style.css">
+        <script src="https://code.getmdl.io/1.3.0/material.min.js"></script>
+        <script src="./script/login.js" defer type="module"></script>
+        <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css"
+            integrity="sha512-SnH5WK+bZxgPHs44uWIX+LLJAJ9/2PkPKZ5QiAj6Ta86w+fsb2TkcmfRyVX3pBnMFcV7oQPJkl9QevSCWr3W6A=="
+            crossorigin="anonymous" referrerpolicy="no-referrer" />
+    
+            <link rel="apple-touch-icon" sizes="180x180" href="/Icon/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/Icon/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/Icon/favicon-16x16.png">
+    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
+    <meta name="msapplication-TileColor" content="#da532c">
+    <meta name="theme-color" content="#ffffff">
+    <link rel="manifest" href="/manifest.json" />
+    </head>
+<body>
+    <header>
+        <div class="app-name">
+            <img class="money-img" src="spenwise-tab.webp" alt="money">
+            <span class="app-name-text">SPEND WISE</span>
+        </div>
+    </header>
+    <div class="category">
+     <div id="category-container" class="category-container"></div>
+     <p id="addBtn" class="category-title">New <i id="addButton" class="fa-solid fa-plus"></i></p>
+    </div>
+    <div id="add-category-dialog" class="floating-dialog">
+        <h4 id="category-dialog-title">Category Name</h4>
+        <h5>Set Limit</h5>
+        <div style="width: 100%;">
+            <input class="mdl-slider mdl-js-slider" type="range" id="s1" min="0" max="1000" value="100">
+        </div>
+        <div>
+            <span>0</span>
+            <span style="float: right;">1000</span>
+        </div>
+        <p id="current-limit"></p>
+        <div class="mdl-dialog__actions">
+            <button type="button" class="mdl-button close-slide-dialog">Cancel</button>
+            <button type="button" class="mdl-button mdl-js-button mdl-button--raised mdl-button--accent"
+                id="add-category-ok">OK</button>
+        </div>
+    </div>
+    <div id="formContainer" class="form-container">
+        <h4 style="font-weight: bold;">Add Category</h4>
+        <div class="form-body">
+            <h6 for="categoryName">Category Name:</h6>
+            <input type="text" class="mdl-textfield__input" id="categoryName"
+                placeholder="Enter category name"><br><br>
+            <div>
+                <button id="submitForm" type="button"
+                    class="mdl-button mdl-js-button mdl-button--raised mdl-button--accent">ADD</button>
+                <button id="closeForm" class="mdl-button close">CLOSE</button>
+            </div>
+        </div>
+    </div>
+    <div class="tabs">
+        <a href="index.html" >Home</a>
+        <a style="border-top: 2px solid rgb(0, 98, 255);" href="category.html">Categories</a>
+        <a href="transactions.html" >Transactions</a>
+        <a href="statistics.html">Statistics</a>
+    </div>
+<script type="module" src="./script/category.js"></script>
+</body>
+</html>
+
+
+
+
+
+
+
