+++ statistics.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Spend Wise</title>
+        <link rel="stylesheet" href="./css/material.min.css">
+        <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+        <link rel="stylesheet" href="css/style.css">
+        <script src="https://code.getmdl.io/1.3.0/material.min.js"></script>
+        <script src="./script/login.js" defer type="module"></script>
+        <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css"
+            integrity="sha512-SnH5WK+bZxgPHs44uWIX+LLJAJ9/2PkPKZ5QiAj6Ta86w+fsb2TkcmfRyVX3pBnMFcV7oQPJkl9QevSCWr3W6A=="
+            crossorigin="anonymous" referrerpolicy="no-referrer" />
+    
+            <link rel="apple-touch-icon" sizes="180x180" href="/Icon/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/Icon/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/Icon/favicon-16x16.png">
+    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
+    <meta name="msapplication-TileColor" content="#da532c">
+    <meta name="theme-color" content="#ffffff">
+    <link rel="manifest" href="/manifest.json" />
+    </head>
+<body>
+  <div>
+    <button id="pie-chart-btn"
+    class="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--accent">
+    Pie Chart
+  </button>
+  <canvas id="pieChartCanvas" style="display: none;"></canvas>
+  <button id="close-pie-chart-btn"
+      class="mdl-button mdl-js-button mdl-button--raised mdl-js-ripple-effect mdl-button--accent"
+      style="display: none; margin: 20px auto;">
+      Close Pie Chart
+  </button>
+</div>
+  <div class="tabs">
+    <a href="index.html" >Home</a>
+    <a href="category.html">Categories</a>
+    <a href="transactions.html" >Transactions</a>
+    <a style="border-top: 2px solid rgb(0, 98, 255);" href="statistics.html">Statistics</a>
+</div>
+  <script type="module" src="./script/statistics.js"></script>
+</body>
+</html>
