+++ transactions.html
@@ -1,0 +1,96 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>Spend Wise</title>
+        <link rel="stylesheet" href="./css/material.min.css">
+        <link rel="stylesheet" href="https://fonts.googleapis.com/icon?family=Material+Icons">
+        <link rel="stylesheet" href="css/style.css">
+        <script src="https://code.getmdl.io/1.3.0/material.min.js"></script>
+        <script src="./script/login.js" defer type="module"></script>
+        <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.2/css/all.min.css"
+            integrity="sha512-SnH5WK+bZxgPHs44uWIX+LLJAJ9/2PkPKZ5QiAj6Ta86w+fsb2TkcmfRyVX3pBnMFcV7oQPJkl9QevSCWr3W6A=="
+            crossorigin="anonymous" referrerpolicy="no-referrer" />
+    
+            <link rel="apple-touch-icon" sizes="180x180" href="/Icon/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/Icon/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/Icon/favicon-16x16.png">
+    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#5bbad5">
+    <meta name="msapplication-TileColor" content="#da532c">
+    <meta name="theme-color" content="#ffffff">
+    <link rel="manifest" href="/manifest.json" />
+    </head>
+<body>
+    <div class="comparison">
+        <div style="display: flex; gap: 25px;">
+        <p id="current-month-total"></p>
+        <p id="previous-month-total"></p>
+        </div>
+        <div>
+        <p id="expenditure-comparison"></p>
+    </div>
+    </div>
+    <div id="expense-list"></div>
+    <div class="fab-container">
+        <button id="fab"
+            class="mdl-button mdl-js-button mdl-button--fab mdl-js-ripple-effect mdl-button--colored">
+            <i class="material-icons">add</i>
+        </button>
+    </div>
+    <dialog class="mdl-dialog">
+        <h4 id="dialog-title" class="mdl-dialog__title">Add Transaction</h4>
+        <div class="mdl-dialog__content">
+            <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                <input class="mdl-textfield__input" type="number" min="0" max="10000000000" id="amount"
+                    required>
+                <label class="mdl-textfield__label" for="amount">Amount</label>
+                <span class="mdl-textfield__error">Input is not a number!</span>
+            </div>
+            <div class="mdl-textfield mdl-js-textfield">
+                <label class="mdl-radio mdl-js-radio mdl-js-ripple-effect" for="credit">
+                    <input type="radio" id="credit" class="mdl-radio__button" name="transactionType"
+                        value="credit" required>
+                    <span class="mdl-radio__label">Credit</span>
+                </label>
+                <label class="mdl-radio mdl-js-radio mdl-js-ripple-effect" for="debit">
+                    <input type="radio" id="debit" class="mdl-radio__button" name="transactionType"
+                        value="debit">
+                    <span class="mdl-radio__label">Debit</span>
+                </label>
+            </div>
+            <div id="category-field"
+                class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label hidden">
+                <select class="mdl-textfield__input" id="category">
+                    <option value=""></option>
+                </select>
+                <label class="mdl-textfield__label" for="category">Category</label>
+            </div>
+            <div class="mdl-textfield mdl-js-textfield">
+                <label class="mdl-checkbox mdl-js-checkbox mdl-js-ripple-effect" for="recurring">
+                    <input type="checkbox" id="recurring" class="mdl-checkbox__input">
+                    <span class="mdl-checkbox__label">Monthly Recurring</span>
+                </label>
+            </div>
+            <div class="mdl-textfield mdl-js-textfield mdl-textfield--floating-label">
+                <input class="mdl-textfield__input" type="text" id="description">
+                <label class="mdl-textfield__label" for="description">Description (optional)</label>
+            </div>
+        </div>
+        <div class="mdl-dialog__actions">
+            <button type="button"
+                class="mdl-button close mdl-js-button mdl-button--raised mdl-js-ripple-effect">Cancel</button>
+            <button type="button"
+                class="mdl-button mdl-js-button mdl-button--raised mdl-button--accent">OK</button>
+        </div>
+    </dialog>
+    <script type="module" src="./script/transactions.js"></script>
+    <div class="tabs">
+        <a href="index.html" >Home</a>
+        <a href="category.html">Categories</a>
+        <a style="border-top: 2px solid rgb(0, 98, 255);" href="transactions.html" >Transactions</a>
+        <a href="statistics.html">Statistics</a>
+    </div>
+</body>
+</html>
